Adding script that assigns people to workplaces and schools
School assignment is based on simple proximity, i.e. students are
sent to the nearest school.

Workplace assignment is probabilistic according to a gravity model
that takes into account the distance to between home and the workplace,
and the number of jobs available at that workplace.  The 1000 nearest
workplaces are sampled from.