Immunity file can have time since infection.
Population immunity file can now list the number of years ago (from the present) a person was infected by each serotype, instead of just 0/1. A value of 0 indicates no prior exposure to the serotype. This new format is basically compatible with the original population immunity files, but it allows for cross-serotype immunity to last for an arbitrary number of years.