Swap immune states on day 100, use only one timer
Swapping immune states on day 0 of each year (January 1) is problematic
because it tends to be during an epidemic, thus causing a sudden decrease
in herd immunity and an uptick in the epidemic trajectory.

Swapping on day 100 is a stop-gap solution, since this is in the middle of
the off-season in the Yucatan. Better would be to swap 1/365 of the people
every day, but this is computationally more challenging and there are some
details that would need to be worked out with keeping track of which locations
have infectious people.

Also, we previously were using two different day counters/timers, which meant
that in different scopes regarded "today" as possibly different days.