bugfixes, mostly related to running the simulator > once in the same process
Some objects constructed with new were never being deleted

getPersonByID() was working the way we were using, it, but it would fail in other
reasonable use cases.

The serial ID counter in Person could not be reset in cases where the same thread
would try to run the simulator a second time.

Moved contents of Utility.h to its own dengue::util namespace to avoid collisions.

Added back the # of mosquitoes as a fitted parameter to ABC program.