Bug fixes with how vaccinations are distributed
MAX_PERSON_AGE was a problematic parameter name, because it didn't actually
mean "maximum person age."  Rather, it was the size of some arrays--which means
that it is not also a valid index, although age is sometimes used as an index.
This resulted in seg faults in some vaccination scenarios.  I've renamed the
parameter to NUM_AGE_CLASSES, because that's principally how it's being used, and
fixed some other weird behavior with vaccinating age groups as well:

 - People could receive boosted (non-naive) immunity if vaccinated a second time
   using a specified negative age, but not if positive age values were specified.
   Aside from issues with all negative ages being non-intuitively processed as
   meaning "vaccinate everyone," the boosting behavior is subtle and idiosyncratic.
   I've modified the code so that no individual can be vaccinated twice.  Booster
   vaccines should be modeled explicitly if we intend to use them.

 - If multiple ages were specified for vaccination that were > the MAX_PERSON_AGE,
   the resulting behavior is that every individual in the oldest age group would
   receive extra opportunities to be vaccinated.  The behavior is idiosyncratic
   and therefore was treated as a bug and has been fixed so that vaccines specified
   for individuals of a certain age that do not exist within a population do not
   get applied to anyone.