Some bugfixes affecting vaccine ABC runs
- There was some weirdness with how RNGs were being handled/seeded,
stemming from the fact that we were using two, both named RNG, and
with nested scopes.  I removed one, so now we're sharing an RNG for
the ABC stuff and the epidemic simulation stuff.  It's not apparent
that this was ever actually causing problems, but it seems like it
could have been.

- VESs_NAIVE was not being set under the all_mature scenario, but it
was being used.  This was a misunderstanding on my part regarding
how the VES values were being used.

- People with age max_catchup_age were not being vaccinated because
the last age being specified for vaccination was max_catchup_age - 1.

- Renamed *.qsub and *.json files to confuse me less, improved stuff
being output to make life easier, changed burn-in period and num
mosquitoes for a low FOI version, maybe some other fiddly things as well.