Minor bugfix affecting bad filename usage
If filenames were specified for things like serotype and EIP data, but
appropriate files do not actually exist and therefore a filehandle cannot be
opened, an error message was sent on stderr, but the program would continue to
run, potentially segfaulting.  Appropriate behavior is to exit gracefully if
the user provides arguments that cannot be used for any reason.