Adding support for generating serotype patterns
Empirical serotype time series are "streaky", meaning serotypes tend to be
observed for several years in a row and then be absent for several years. These
runs and gaps appear to be approximately geometrically distributed.  We now
support randomly generating serotype introduction time series.  Currently the
geometric distributions are parameterized based on the serotype data we observe
for Yucatan.