Adding utility function for randomizing dynamic parameters
The function is generic, but the current use case is a daily-varying
parameter that we want to randomize annually, so that the sequence
stays in tact within each year, but the years occur in different orders
between simulation runs.  This is so that, when averaging many runs,
trends due to common, dynamic parameters do not produce distinct temporal
trends in the output that would confuse readers in certain contexts.