Adding support to specify day of calendar year to start simulation
Also: Swapping now happens on last day of simulation year, and vaccination
happens on first day of simulation year.  Monthly reporting is done according
to the standard calendar, but yearly reporting is done according to the
simulation year, i.e. after 365 simulated days, the first year of data is
reported, regardless of whether it coincides with December 31.

Adding ten more years to the burn-in.  Now we're up to a 50-year burn-in, of
which the first 30 years aren't output.