bugfix affecting initial immunity
When loading initial immunity, we were not handing infection history and
immunity very consistently, e.g. individuals could have past infections
with recovery times but no infection times.  This seems to have caused a
bug, the details of which I have not teased apart, but in any case the way
we were handling infection histories was not consistent, and making it
consistent made the bug (immunity decreased the older individuals were
at the beginning of the simulation) go away.

Some details:

- immune state changes due to infections, past or present, should only be
made by calling Person::infect, and not by tinkering around with their immune
state directly

- new constructor taking serotype for Infection class, which is the constructor
that should generally be used

- some refactoring of Community::loadPopulation, specifically the parser for
population immunity files

- new function to output current population immunity in simulator.h.  this
could use a little bit of refining, specifically it would be nice if a filename
could be passed in as an argument, and the times output are relative to the
time at the end of the simulation, even if they are being output at a different
time.

- new way of handling daily output.  previous approach used cout, which is not
viable for paralellized runs.  now we store output in a vector during the run,
and flush it all at once to a file when the run is over.