bugfix; substantial speed improvements
About 0.1% of locations do not have neighbors (i.e. degree == 0) to which
mosquitoes could move to.  If an infected mosquito tried to move from a
neighborless location, a segfault would occur.  Since uninfected mosquitoes
are not explicitly modeled, this did not affect them.  The scenario where
the segfault would occur appears to be very uncommon.

This commit also changes a map that was tracking locations with viremic people
to a vector of unsorted sets, and results in a ~30% improvement in running
time.