Changes to timing of vaccination, boosting, and seroprevalence measurement
Vaccination used to always be on day one of the simulator; for now it's
hard-coded as Julian day 100, because day 100 better approximates the
middle of the interepidemic period. We should add a parameters attribute for this.

Seroprevalence should have been taken around Julian day 100, and instead
was synchned with the end of the simulation year.

Boosting is now set to occur every two years if there is waning, per Ira.