Critical bugfix affecting reading immunity files; some minor fixes
Infections are not allowed to happen out of chronological order; this normally
is a reasonable restriction on the model, but in the immunity file format,
infections are ordered by serotype, rather than by date of infection.  When
parsing, out of order infections were being silently ignored, resulting in the
population having less immunity than it should have according to the immunity
file.

Additional things:
This commit fixes an invalid person ID in the toy swap file affecting person
99999.

There were some inconsistent checks when getting someone by ID.  For now, ID's
start at 1, not 0, so trying to get the person with ID 0 should be an error.

Finally, when checking whether someone is infectable, crossimmunity is
considered from the date of infection, rather than the date of recovery.  This
may be slightly less biologically sensible, but it's much more straightforward
to handle computationally, and frankly, the precise nature of crossimmunity for
dengue is too fuzzy for this to make a difference.