Supporting loading a daily mos multiplier file; changed run duration
The major changes are related to reading in a daily mosquito multiplier
file.  Also changed how DDT is modeled--introductions AND mosquito populations
are reduced.

Reduced the total simulated years from 155 to 135.  The initial burn-in period
was longer than it needed to be, based on the ~60 years of consistent epidemics
before the DDT era.

Also fixed a seasonality figure, which was using eip(avg(daily_tmp)) instead of
avg(eip(daily_tmp)), which we believe may be a better approximation.