Refactored Person::infect; new sampler ftn; new abc pars
We used to use "secondary" to mean "post-primary," which lead to confusion.
Pathogenicities and probabilities of severe disease given disease are now fully
stratified by number of past infections.  We're getting rid of the ability to
express secondary pathogenicity as an odds ratio, since it's not clear how we
would want to express tertiary and quaternary pathogenicities.  It's much
simpler to express them as independent pars, and leave it to the user to do
more complicated transformations.

Our sampling from CDFs was a bit clunky and repetitious, so we're adding a
generic "broken stick" sampler in Parameters that is static and thus accessible
from anywhere.

We're no longer supporting setting a maximum number of infections that is not
equal to the number of serotypes circulating (max parity).  We don't use this,
and it's not consistent with what we know from serosurvey data.

Making more arrays into vectors.

Refactored Person::infect for clarity and to use the new pathogenicity/severity
parameters.

Minor additions to fitting notes.

Updates to ABC code to use the new parameters.