Changed enum names for clarity/consistency; corrected defaults
The last commit (and this one) compile without warnings using gcc 4.7.2.  I think
there may be a clang-specific issue.

WHO vaccine mechanisms that differ from our normal assumptions should be made
in the main program in the who_proj directory, rather than via the Parameters
constructor.

bRetroactiveMatureVaccine was not being set anywhere, but all data members
should be explicitly set in the constructor (which means in this case
define_defaults()).  As it was implicitly defaulting to false, we are now
explicitly doing the same.