Adding support for more complicated climate model
We have fairly good temperature and precipitation data for the fitting period,
but not much before that.  We're now assuming that historical weather had only
intra-annual variation, and was the average of the first 10 years of the fitted
period, while every year in the fitted period is unique.  We've made
modifications to support reading in multiple files for EIP and mosquito
multiplier values.