Adding serial to output; updates to abc and vaccine runs
This commit seems to include a bugfix to vaccine_abc/main_sql.cpp,
which previously was producing epidemics that were too large.  It's
likely that slightly different models where being specified for abc
and vaccine runs, and that was driving the difference.  After diffing
the two files and addressing differences, the difference in epidemic
sizes seems to have been resolved as well.