Switching to set from unordered_set
We've observed some instances in which simulations produce inconsistent
output, even when seeded the same way.  I previously tracked this down to
the order with which locations are iterated through in the unordered isHot
set.  It's possible, for example, that unordered sets of Location* were hashed
based on memory address, and the order would change when different hardware
was used.  Switching to and (ordered) set might result in slower simulations,
but if so, they are not obviously slower in small scale tests.