A bugfix and many improvements to IRS vector control
Major bugfix: location type was not being set, thus meaningless values
were being used when checking whether a loc was a house.

Each loc now has a priority queue of VC treatments, so they can all be
queued up at the beginning of the simulation and we don't have to worry
about e.g. whether they span calendar years.

VC is now updated and infected mosquitoes affected at the beginning of
the day, rather than the end.  This is not important, except that it
makes the code easier to interpret--fewer tweaks needed to get day
accounting right.

Aggregating of infections/cases into years needs to be based on when
IRS treatment actually started for timing analysis, so that we are
not unfairly penalizing IRS that starts late in the calendar year.