Refitting model; new metrics and new parameterization
We found some errors in the empirical data we'd used before, and that we had been incorrectly
aggregating epidenics on a Yucatan epidemic calendar, rather than the standard (if illogical)
Julian calendar.  Also, we have a couple more years of data, so we're refitting.

Parameters are sampled in a transformed (logit) space now, using gaussian priors, and noising
is done after taking into account covarainces between parameters.

DB tables are now job, met, par, and upar, with upar containing the untransformed, model-space
parameters.