Bugfix in how students were sent to school
Students were in many cases being sent to the closest workplace or school,
rather than strictly the closest school.  The Merida subpopulation needs
to be updated.  population-yucatan.txt has been updated on my machines and
is not in the repository at this point (it's 68MB).  TODO - decide on a way
to distribute the pop file.  I think the old version may be bundled in the
PLoS NTDs release that we put together.