Added (commented-out) support for R0 sensitivity analysis
For a normal R0 run, we're running the simulator (with replicates) for every day of the year,
for the entire parameter posterior distribution.

Carl and I wanted to know what the relative contributions of mosquito population size and EIP
were to R0, which means fixing the other parameters and doing a 2-D sweep of those two variables.