Significant bugfix affecting vaccine scenarios
When we changed the aging model to allow birthdays throughout the year,
rather than on the same day each year, that had complications for the
vaccination model, which was implemented for people of specific ages on
a particular day of the year.

Typical example:  If 9-year-olds were vaccinated on Julian day 100, anyone
turning 9 afterward would not be vaccinated until the following year's day
100.  Some people turning age 10 would inherit their immune state from
some of these unvaccinated 9-year-olds, however.  Since they are no longer
the target age, these new unvaccinated people age 10 or older would never
be vaccinated.

The solution I've adopted is to still have a one day, mass catchup campaign
on the specified day, but routine vaccinations happen on birthdays.

Also in this commit, I've removed some old code that no longer served a
purpose.