Bugfixes to vaccine models, still working on validation
Boosting and multi-doses vaccines weren't happening unless they fell on recipient's
birthday.  We're now using a revaccination set to determine which people might need
another dose, and we check it every day.  Because the re-vac interval is possibly
complicated (particularly if people are both getting multiple doses and bossted, and
not on a yearly schedule), it seems safer for now to check everyone who might need
re-vaccination, and not try to do the bookkeeping of who is due on a particular day.
Revisiting that would potentially substantially speed up runs where many people get
many vaccines.

After changing the routine vaccination approach, so that now people are vaccinated on
their birthdays, we were vaccinating from the start of the simulation, rather than
waiting until the intented intervention year.

I noticed that with the latest gcc compiler available on hipergator, partial_sum has
been moved from algorithm to numeric, resulting in surprising compile errors.  We're
now incuding both so that it doesn't matter.