New mechanism to give CYD-TDV to perceived seropositives only
I.e. using imperfect testing.  I believe this commit also addresses
some old inconsistencies with tallying, where the starting and ending
dates used when aggregating years varyied depending on the intervention
used.  This made sense when we were sweeping across the year with TIRS,
but created artifacts in the results when intervention timing is fixed.