Bugfixes and several improvements
Prevalence reported in the log file wasn't correct, but afaik it was never
used for anything.

Standardized the order of arguments to Person::infect and
Person::initializeNewInfection.

Changed some arguments that were IDs to object pointers instead.  This is
clearer, more robust, and more useful.  It also allows the compiler to do
its damn job.

Added an enum to represent trial arm state.