Implementing infection history data pipeline
Copied from COVID-ABM.

Added gen_infection_hist_db to Parameters to control whether the
data is dumped to a database or not.

Added new data members to Infection: infectionOwner.

Added more getter methods to Infetion.

Added generate_infection_db() function to simulator.h. This function
handles the generation of temporary CSV files and SQL scripts to be
used in the creation and population of the final database. This
function then deletes these temporary files so that the only result
is the populated infection database.